// semFront/src/app/models/estudio-interface.ts
export interface EstudioInterface{
    id?: any;
    area?: any;
    categoria?: any;
    domingos?: any;
    sexo?: any;
    sala?: any;
    prueba?: any;
    sinAcro?: any;
    beneficio?: any;
    tipoMuestra?: any;
    indicacionesPreExamen?: any;
    otrasIndicaciones?: any;
    tiempoEntrega?: any;
    precioVenta?: any;
    precioApoyo?: any;
    precioNormal?: any;
}