// semFront/src/app/models/tiempo-interface.ts
export interface TiempoInterface{
    results?: any;
    id?: string;
    Analisis?: string;
    US?: string;
    RXS?: string;
    RXC?: string;
    Masto?: string;
    TACS?: string;
    TACC?: string;
    RMS?: string;
    RMC?: string;
    Densi?: string;
    EKG?: string;
    Papa?: string;
    Colpo?: string;
    Clinica?: string;
    Audio?: string;
    Espiro?: string;
    Pato?: string;
}